requirements.txt
discord.py>=0.16.12